+++ tor_brute.py
@@ -41,22 +41,21 @@ if sys.version_info < (3, 8):
 
 
 import logging
-from stem.util.log import get_logger
 
-from termcolor import colored
 from alive_progress import config_handler
 from alive_progress.animations.bars import bar_factory
 from alive_progress.animations.spinners import frame_spinner_factory
+from stem.util.log import get_logger
+from termcolor import colored
 
+from src.core import App, check_args, parser
 from src.helper.logger import init_logger
-from src.core.cli_parser import parser, check_args
-from src.core.app import App
-
 
 if __name__ == "__main__":
-    args = check_args(parser())
+    args = check_args(parser().parse_args())
     supports_color = init_logger(logging.DEBUG if args.debug else logging.INFO)
 
+    # https://stackoverflow.com/a/46400076/13708995
     stem_logger = get_logger()
     stem_logger.propagate = False